fix install_ddrig_deps script (#15)
changes the install_ddrig_deps.sh to instead call _install_ddrig_deps.R, as running an actual R script is a cleaner, smoother way to handle running R from bash. the .R file is prefixed with an _ which is standard practice for implying that something (whether it's a file in a repo or a function in code) is meant for internal use.

Anyone who wants to pull the DDRIG repo can now simply call ./scripts/install_ddrig_deps.sh and all libraries will be installed within a matter of about 15 minutes (if the user has none of the libraries installed at all).